requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pymysql==1.1.0
cryptography==41.0.7